# BACKEND/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rto": {
      "get": {
        "operationId": "RtoController_getAllRtos",
        "parameters": [
          {
            "name": "warehouseId",
            "required": false,
            "in": "query",
            "description": "Filter by warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all RTOs (optionally filtered by warehouse)",
        "tags": [
          "RTO"
        ]
      }
    },
...
    "/su/overview": {
      "get": {
        "operationId": "SuperUserController_getOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get system overview (counts, stats)",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/admins": {
      "get": {
        "operationId": "SuperUserController_getAdmins",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all admin nodes",
        "tags": [
          "Super User"
        ]
      },
      "post": {
        "operationId": "SuperUserController_createAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a new admin node",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/admins/{id}": {
      "put": {
        "operationId": "SuperUserController_updateAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update an admin node",
        "tags": [
          "Super User"
        ]
      },
      "delete": {
        "operationId": "SuperUserController_deleteAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete an admin node",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/admins/{id}/status": {
      "put": {
        "operationId": "SuperUserController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAdminStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Set active/inactive status for an admin",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/warehouses": {
      "get": {
        "operationId": "SuperUserController_getWarehouses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all warehouses",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/warehouses/{id}": {
      "get": {
        "operationId": "SuperUserController_getWarehouseDrilldown",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get drilldown details for a warehouse",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/warehouses/{id}/pending": {
      "delete": {
        "operationId": "SuperUserController_cancelAllPending",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Cancel all pending shipments for a warehouse (bypass)",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/warehouses/{id}/pending/{orderId}": {
      "delete": {
        "operationId": "SuperUserController_cancelSinglePending",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "Order ID to cancel",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Cancel a specific pending shipment for a warehouse",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/hubs": {
      "get": {
        "operationId": "SuperUserController_getHubs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all transit hubs",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/hubs/{id}": {
      "get": {
        "operationId": "SuperUserController_getHubDrilldown",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get drilldown details for a transit hub",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/hubs/{id}/outscan-all": {
      "put": {
        "operationId": "SuperUserController_outscanAllHubShipments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Force out-scan all shipments from a transit hub (bypass)",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/agencies": {
      "get": {
        "operationId": "SuperUserController_getAgencies",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all local agencies",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/agencies/{id}": {
      "get": {
        "operationId": "SuperUserController_getAgencyDrilldown",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get drilldown details for an agency",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/agencies/{id}/agents/{agentId}/status": {
      "put": {
        "operationId": "SuperUserController_updateAgentStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "Active",
                    "enum": [
                      "Active",
                      "Busy",
                      "Offline"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update the status of an agent within an agency",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/shipments": {
      "get": {
        "operationId": "SuperUserController_getShipments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all shipments",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/rto": {
      "get": {
        "operationId": "SuperUserController_getRTO",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all RTOs",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/tickets": {
      "get": {
        "operationId": "SuperUserController_getTickets",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all tickets",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/logs": {
      "get": {
        "operationId": "SuperUserController_getLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get system activity logs",
        "tags": [
          "Super User"
        ]
      },
      "post": {
        "operationId": "SuperUserController_appendLog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Append a new log entry",
        "tags": [
          "Super User"
        ]
      }
    },
    "/su/config": {
      "get": {
        "operationId": "SuperUserController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get system configuration",
        "tags": [
          "Super User"
        ]
      },
      "put": {
        "operationId": "SuperUserController_setConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update system configuration (thresholds, limits)",
        "tags": [
          "Super User"
        ]
      }
    }
  },
  "info": {
    "title": "WARE2DOOR API",
    "description": "## WARE2DOOR Logistics API\n\nEnd-to-end logistics platform covering warehouse management, transit hub scanning, last-mile delivery, RTO lifecycle, and super-user administration.\n\n### Authentication\nProtected endpoints require an **`x-role`** header specifying the caller's role.\n\nClick the **Authorize 🔒** button, then enter one of the allowed values:\n\n| Role Value | Access |\n|---|---|\n| `WAREHOUSE` | Warehouse management endpoints |\n| `TRANSIT_HUB` | Transit hub scan & reports endpoints |\n| `LOCAL_AGENCY` | Agency delivery & agent endpoints |\n| `SUPER_USER` | Full administrative access |\n\n> **Public endpoints** (login, verify-email, track, raise-ticket) do **not** require this header.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role-based access header. Allowed values: WAREHOUSE | TRANSIT_HUB | LOCAL_AGENCY | SUPER_USER"
      }
    },
    "schemas": {
...
      "CreateAdminDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "WareHouse",
            "enum": [
              "WareHouse",
              "TransitHub",
              "LocalAgency"
            ],
            "description": "Role of the admin node"
          },
          "name": {
            "type": "string",
            "example": "Hyderabad Warehouse",
            "description": "Name of the admin/facility (3-80 chars)"
          },
          "email": {
            "type": "string",
            "example": "hyd.warehouse@ware2door.com",
            "description": "Valid email address"
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Valid 10-digit Indian mobile number"
          },
          "facility": {
            "type": "string",
            "example": "Hyderabad Central Warehouse",
            "description": "Facility name (optional)"
          },
          "city": {
            "type": "string",
            "example": "Hyderabad",
            "description": "City where the node is located"
          },
          "password": {
            "type": "string",
            "example": "Admin@123",
            "description": "Initial password (6-64 chars, optional)"
          }
        },
        "required": [
          "role",
          "name",
          "email",
          "phone",
          "city"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Hyderabad Warehouse",
            "description": "Updated name (3-80 chars)"
          },
          "email": {
            "type": "string",
            "example": "new@ware2door.com",
            "description": "Updated email address"
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Updated 10-digit Indian mobile number"
          },
          "city": {
            "type": "string",
            "example": "Mumbai",
            "description": "Updated city"
          }
        }
      },
      "SetAdminStatusDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "example": true,
            "description": "Set to true to activate, false to deactivate the admin"
          }
        },
        "required": [
          "active"
        ]
      },
      "AppendLogDto": {
        "type": "object",
        "properties": {
          "admin": {
            "type": "string",
            "example": "Super Admin",
            "description": "Name of the admin performing the action (max 80 chars)"
          },
          "module": {
            "type": "string",
            "example": "Shipments",
            "description": "Module where the action was performed (max 60 chars)"
          },
          "action": {
            "type": "string",
            "example": "Cancel All Pending",
            "description": "Action that was performed (max 100 chars)"
          },
          "details": {
            "type": "string",
            "example": "Cancelled all pending shipments for WH-HYD-001",
            "description": "Detailed description of the action (max 500 chars)"
          }
        },
        "required": [
          "admin",
          "module",
          "action",
          "details"
        ]
      },
      "SetConfigDto": {
        "type": "object",
        "properties": {
          "lowStockThreshold": {
            "type": "number",
            "example": 50,
            "description": "Threshold below which stock is flagged as low (1-10000)"
          },
          "maxRtoAttempts": {
            "type": "number",
            "example": 3,
            "description": "Maximum delivery attempts before initiating RTO (1-10)"
          }
        }
      }
    }
  },
  "security": [
    {
      "x-role": []
    }
  ]
}